+++ packages/core/lib/registry/registry.js
@@ -21,9 +21,10 @@ const createRegistry = () => {
   const registry = {
     init (broker, middlewareHandler, serviceChanged) {
       this.broker = broker
-      // this.transport = transport
+
       this.middlewareHandler = middlewareHandler
       this.serviceChanged = serviceChanged
+
       // create logger
       this.log = broker.createLogger('REGISTRY')
 
@@ -42,11 +43,11 @@ const createRegistry = () => {
     onRegisterLocalAction: noop,
     onRegisterRemoteAction: noop,
     /**
-         * Check action visibility
-         * @param {*} action Action definition
-         * @param {*} node Node object
-         * @returns {Boolean} Is visible
-         */
+     * Check action visibility
+     * @param {*} action Action definition
+     * @param {*} node Node object
+     * @returns {Boolean} Is visible
+    */
     checkActionVisibility (action, node) {
       if (typeof action.visibility === 'undefined' || action.visibility === 'public') {
         return true
@@ -57,11 +58,11 @@ const createRegistry = () => {
       return false
     },
     /**
-         *
-         * Register a local service
-         * @param {*} svc Service definition
-         * @returns {void}
-         */
+     *
+     * Register a local service
+     * @param {*} svc Service definition
+     * @returns {void}
+    */
     registerLocalService (svc) {
       if (!this.services.has(svc.name, svc.version, this.broker.nodeId)) {
         const service = this.services.add(this.nodes.localNode, svc.name, svc.version, svc.settings)
@@ -225,22 +226,25 @@ const createRegistry = () => {
       } else {
         if (opts && opts.nodeId) { // remote
           const endpoint = this.getActionEndpointByNodeId(actionName, opts.nodeId)
+
           if (!endpoint) {
-            this.log.warn(`Service ${actionName} is not registered on node ${opts.nodeId}.`)
+            this.log.warn(`Service "${actionName}" is not registered on node ${opts.nodeId}.`)
             return new WeaveServiceNotFoundError({ actionName, nodeId: opts.nodeId })
           }
+
           return endpoint
         } else {
           const endpointList = this.getActionEndpoints(actionName)
+
           if (!endpointList) {
-            this.log.warn(`Service ${actionName} is not registered.`)
+            this.log.warn(`Service "${actionName}" is not registered.`)
             return new WeaveServiceNotFoundError({ actionName })
           }
 
           const endpoint = endpointList.getNextAvailable()
 
           if (!endpoint) {
-            this.log.warn(`Service ${actionName} is not available.`)
+            this.log.warn(`Service "${actionName}" is not available.`)
             return new WeaveServiceNotAvailableError({ actionName })
           }
 
@@ -265,18 +269,27 @@ const createRegistry = () => {
       const endpointList = this.getActionEndpoints(actionName)
 
       if (!endpointList) {
-        this.log.warn(`Service ${actionName} is not registered localy.`)
+        this.log.warn(`Service "${actionName}" is not registered localy.`)
         throw new WeaveServiceNotFoundError({ actionName })
       }
 
       const endpoint = endpointList.getNextLocalEndpoint()
 
       if (!endpoint) {
-        this.log.warn(`Service ${actionName} is not available localy.`)
+        this.log.warn(`Service "${actionName}" is not available localy.`)
         throw new WeaveServiceNotAvailableError({ actionName })
       }
 
       return endpoint
+    },
+    getNodeInfo (nodeId) {
+      const node = this.nodes.get(nodeId)
+
+      if (!node) {
+        return null
+      }
+
+      return node.info
     },
     getLocalNodeInfo (forceGenerateInfo) {
       if (forceGenerateInfo || !this.nodes.localNode.info) {
@@ -285,7 +298,7 @@ const createRegistry = () => {
 
       return this.nodes.localNode.info
     },
-    generateLocalNodeInfo (incrementSequence) {
+    generateLocalNodeInfo (incrementSequence = false) {
       const { client, IPList, sequence } = this.nodes.localNode
       const nodeInfo = { client, IPList, sequence }
 
@@ -335,13 +348,13 @@ const createRegistry = () => {
 
       if (isNew) {
         this.broker.broadcastLocal('$node.connected', { node, isReconnected })
-        this.log.debug(`Node ${node.id} connected!`)
+        this.log.info(`Node "${node.id}" connected!`)
       } else if (isReconnected) {
         this.broker.broadcastLocal('$node.connected', { node, isReconnected })
-        this.log.debug(`Node ${node.id} reconnected!`)
+        this.log.info(`Node "${node.id}" reconnected!`)
       } else {
         this.broker.broadcastLocal('$node.updated', { node, isReconnected })
-        this.log.debug(`Node ${node.id} updated!`)
+        this.log.info(`Node "${node.id}" updated!`)
       }
     },
     nodeDisconnected (nodeId, isUnexpected) {
@@ -350,13 +363,13 @@ const createRegistry = () => {
         this.deregisterServiceByNodeId(node.id)
         node.disconnected(isUnexpected)
         this.broker.broadcastLocal('$node.disconnected', { nodeId, isUnexpected })
-        this.log.warn(`Node '${node.id}'${isUnexpected ? ' unexpectedly' : ''} disconnected.`)
+        this.log.warn(`Node "${node.id}"${isUnexpected ? ' unexpectedly' : ''} disconnected.`)
       }
     },
     removeNode (nodeId) {
       this.nodes.remove(nodeId)
       this.broker.broadcastLocal('$node.removed', { nodeId })
-      this.log.warn(`Node '${nodeId}' removed.`)
+      this.log.warn(`Node "${nodeId}" removed.`)
     },
     getNodeList (options) {
       return this.nodes.list(options)