+++ packages/core/lib/transport/adapters/tcp/tcpReader.js
@@ -4,7 +4,7 @@ const TCPWriteStream = require('./tcpWriteStream')
 
 module.exports = (adapter, options) => {
   const self = Object.assign({}, EventEmitter.prototype)
-  const sockets = []
+  let sockets = []
   let server
 
   self.isConnected = false
@@ -22,21 +22,46 @@ module.exports = (adapter, options) => {
         const port = server.address().port
         self.isConnected = true
         adapter.log.info(`TCP server is listening on port ${port}`)
-        resolve(options.port)
+        resolve(port)
       })
     })
+  }
+
+  self.close = () => {
+    if (server && self.isConnected) {
+      sockets.forEach(socket => socket.destroy())
+      sockets = []
+    }
   }
 
   function onTCPClientConnected (socket) {
     sockets.push(socket)
 
-    // const address = socket.remoteAddress
     const parser = new TCPWriteStream(adapter, socket)
     socket.pipe(parser)
 
-    parser.on('data', (type, message) => {
-      adapter.onIncomingMessage(type, message)
+    parser.on('error', error => {
+      adapter.log.warn('Packet parser error!', error)
+      closeSocket(socket)
     })
+
+    parser.on('data', (type, message) => {
+      self.emit('message', type, message)
+    })
+
+    socket.on('error', error => {
+      adapter.log.warn('TCP connection error!', error)
+      closeSocket(socket)
+    })
+
+    socket.on('close', (isError) => {
+      closeSocket(socket)
+    })
+  }
+
+  function closeSocket (socket) {
+    socket.destroy()
+    sockets.splice(sockets.indexOf(socket), 1)
   }
 
   return self